chore(engine): move deployment distribution to TransformDeploymentCreateProcessor
