+++ engine/src/main/java/io/zeebe/engine/processing/deployment/TransformingDeploymentCreateProcessor.java
@@ -13,9 +13,12 @@ import io.zeebe.engine.processing.common.CatchEventBehavior;
 import io.zeebe.engine.processing.common.ExpressionProcessor;
 import io.zeebe.engine.processing.common.ExpressionProcessor.EvaluationException;
 import io.zeebe.engine.processing.common.Failure;
+import io.zeebe.engine.processing.deployment.distribute.DeploymentDistributor;
+import io.zeebe.engine.processing.deployment.distribute.PendingDeploymentDistribution;
 import io.zeebe.engine.processing.deployment.model.element.ExecutableCatchEventElement;
 import io.zeebe.engine.processing.deployment.model.element.ExecutableStartEvent;
 import io.zeebe.engine.processing.deployment.transform.DeploymentTransformer;
+import io.zeebe.engine.processing.streamprocessor.ReadonlyProcessingContext;
 import io.zeebe.engine.processing.streamprocessor.TypedRecord;
 import io.zeebe.engine.processing.streamprocessor.TypedRecordProcessor;
 import io.zeebe.engine.processing.streamprocessor.sideeffect.SideEffectProducer;
@@ -27,23 +30,31 @@ import io.zeebe.engine.state.KeyGenerator;
 import io.zeebe.engine.state.ZeebeState;
 import io.zeebe.engine.state.immutable.TimerInstanceState;
 import io.zeebe.engine.state.instance.TimerInstance;
+import io.zeebe.engine.state.mutable.MutableDeploymentState;
 import io.zeebe.engine.state.mutable.MutableEventScopeInstanceState;
 import io.zeebe.engine.state.mutable.MutableWorkflowState;
 import io.zeebe.model.bpmn.util.time.Timer;
 import io.zeebe.protocol.Protocol;
+import io.zeebe.protocol.impl.record.RecordMetadata;
 import io.zeebe.protocol.impl.record.value.deployment.DeploymentDistributionRecord;
 import io.zeebe.protocol.impl.record.value.deployment.DeploymentRecord;
 import io.zeebe.protocol.impl.record.value.deployment.WorkflowRecord;
+import io.zeebe.protocol.record.RecordType;
 import io.zeebe.protocol.record.RejectionType;
+import io.zeebe.protocol.record.ValueType;
 import io.zeebe.protocol.record.intent.DeploymentDistributionIntent;
 import io.zeebe.protocol.record.intent.DeploymentIntent;
 import io.zeebe.util.Either;
+import io.zeebe.util.sched.ActorControl;
+import io.zeebe.util.sched.future.ActorFuture;
 import java.util.Collections;
 import java.util.List;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import org.agrona.DirectBuffer;
+import org.agrona.ExpandableArrayBuffer;
+import org.agrona.concurrent.UnsafeBuffer;
 
 public final class TransformingDeploymentCreateProcessor
     implements TypedRecordProcessor<DeploymentRecord> {
@@ -60,14 +71,18 @@ public final class TransformingDeploymentCreateProcessor
   private final StateWriter stateWriter;
   private final List<Integer> partitions;
   private final DeploymentDistributionRecord deploymentDistributionRecord;
-  //  private final DeploymentDistributor deploymentDistributor;
+  private final MutableDeploymentState deploymentState;
+  private final DeploymentDistributor deploymentDistributor;
+  private final ActorControl actor;
 
   public TransformingDeploymentCreateProcessor(
       final ZeebeState zeebeState,
       final CatchEventBehavior catchEventBehavior,
       final ExpressionProcessor expressionProcessor,
       final int partitionsCount,
-      final Writers writers) {
+      final Writers writers,
+      final ActorControl actor,
+      final DeploymentDistributor deploymentDistributor) {
     workflowState = zeebeState.getWorkflowState();
     eventScopeInstanceState = zeebeState.getEventScopeInstanceState();
     timerInstanceState = zeebeState.getTimerState();
@@ -76,6 +91,9 @@ public final class TransformingDeploymentCreateProcessor
     deploymentTransformer = new DeploymentTransformer(stateWriter, zeebeState, expressionProcessor);
     this.catchEventBehavior = catchEventBehavior;
     this.expressionProcessor = expressionProcessor;
+    deploymentState = zeebeState.getDeploymentState();
+    this.deploymentDistributor = deploymentDistributor;
+    this.actor = actor;
 
     // partitions
     partitions =
@@ -87,7 +105,25 @@ public final class TransformingDeploymentCreateProcessor
   }
 
   @Override
+  public void onRecovered(final ReadonlyProcessingContext context) {
+    actor.submit(() -> reprocessPendingDeployments(context.getLogStreamWriter()));
+  }
+
+  private void reprocessPendingDeployments(final TypedStreamWriter logStreamWriter) {
+    deploymentState.foreachPending(
+        ((pendingDeploymentDistribution, key) -> {
+          final ExpandableArrayBuffer buffer = new ExpandableArrayBuffer();
+          final DirectBuffer deployment = pendingDeploymentDistribution.getDeployment();
+          buffer.putBytes(0, deployment, 0, deployment.capacity());
+
+          distributeDeployment(
+              key, pendingDeploymentDistribution.getSourcePosition(), buffer, logStreamWriter);
+        }));
+  }
+
+  @Override
   public void processRecord(
+      final long position,
       final TypedRecord<DeploymentRecord> command,
       final TypedResponseWriter responseWriter,
       final TypedStreamWriter streamWriter,
@@ -119,6 +155,25 @@ public final class TransformingDeploymentCreateProcessor
             // todo(zell): push deployment to other partition
             //            deploymentDistributor.pushDeployment(key, partitionId, deploymentEvent);
           });
+
+      // todo(zell): simplify for https://github.com/zeebe-io/zeebe/issues/6173
+      // DEPLOYMENT DISTRIBUTION moved from distribute
+      final ExpandableArrayBuffer buffer = new ExpandableArrayBuffer();
+      deploymentEvent.write(buffer, 0);
+
+      final DirectBuffer bufferView = new UnsafeBuffer();
+      bufferView.wrap(buffer, 0, deploymentEvent.getLength());
+
+      // don't distribute the deployment on reprocessing because it may be already distributed
+      // (#3124)
+      // and it would lead to false-positives in the reprocessing issue detection (#5688)
+      // instead, distribute the pending deployments after the reprocessing
+      sideEffect.accept(
+          () -> {
+            distributeDeployment(key, position, bufferView, streamWriter);
+            return true;
+          });
+
     } else {
       responseWriter.writeRejectionOnCommand(
           command,
@@ -129,6 +184,50 @@ public final class TransformingDeploymentCreateProcessor
           deploymentTransformer.getRejectionType(),
           deploymentTransformer.getRejectionReason());
     }
+  }
+
+  private void distributeDeployment(
+      final long key,
+      final long position,
+      final DirectBuffer buffer,
+      final TypedStreamWriter logStreamWriter) {
+    final ActorFuture<Void> pushDeployment =
+        deploymentDistributor.pushDeployment(key, position, buffer);
+
+    actor.runOnCompletion(
+        pushDeployment, (aVoid, throwable) -> writeCreatingDeploymentCommand(logStreamWriter, key));
+  }
+
+  private void writeCreatingDeploymentCommand(
+      final TypedStreamWriter logStreamWriter, final long deploymentKey) {
+    final PendingDeploymentDistribution pendingDeploymentDistribution =
+        deploymentDistributor.removePendingDeployment(deploymentKey);
+    final DirectBuffer buffer = pendingDeploymentDistribution.getDeployment();
+    final long sourcePosition = pendingDeploymentDistribution.getSourcePosition();
+
+    final DeploymentRecord deploymentRecord = new DeploymentRecord();
+    deploymentRecord.wrap(buffer);
+    final RecordMetadata recordMetadata = new RecordMetadata();
+    recordMetadata
+        .intent(DeploymentIntent.DISTRIBUTED)
+        .valueType(ValueType.DEPLOYMENT)
+        .recordType(RecordType.EVENT);
+
+    actor.runUntilDone(
+        () -> {
+          // we can re-use the write because we are running in the same actor/thread
+          logStreamWriter.reset();
+          logStreamWriter.configureSourceContext(sourcePosition);
+          logStreamWriter.appendFollowUpEvent(
+              deploymentKey, DeploymentIntent.DISTRIBUTED, deploymentRecord);
+
+          final long position = logStreamWriter.flush();
+          if (position < 0) {
+            actor.yield();
+          } else {
+            actor.done();
+          }
+        });
   }
 
   private void createTimerIfTimerStartEvent(
